Fix Firefox bug for images within inline-block
In Firefox, an image styled `max-width:100%` within an inline-block will
display at its default size, and not limit its width to 100% of an
ancestral container. Adding `max-width:100%` to the inline-block utility
protects against this bug. Not adding `!important` because it's not
necessary to meet the purpose of the utility, and some components might
need to set a different max-width.

Fix gh-2